requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
boto3==1.34.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
requests==2.31.0
Pillow==10.1.0
opencv-python==4.8.1.78
moviepy==1.0.3
numpy>=1.26.0
python-multipart==0.0.6
gunicorn==21.2.0